+++ demos/webpack-app/src/playground/machine/activity-set.ts
@@ -4,5 +4,13 @@ import { loopActivity } from './activities/loop-activity';
 import { logActivity } from './activities/log-activity';
 import { ifActivity } from './activities/if-activity';
 import { calculateActivity } from './activities/calculate-activity';
+import { convertValueActivity } from './activities/convert-value-activity';
 
-export const activitySet = createActivitySet([calculateActivity, ifActivity, logActivity, loopActivity, setStringValueActivity]);
+export const activitySet = createActivitySet([
+	calculateActivity,
+	convertValueActivity,
+	ifActivity,
+	logActivity,
+	loopActivity,
+	setStringValueActivity
+]);