+++ editor/src/core/variable-name-formatter.spec.ts
@@ -1,7 +1,13 @@
-import { formatVariableName } from './variable-name-formatter';
+import { formatVariableName, formatVariableNameWithType } from './variable-name-formatter';
 
 describe('formatVariableName', () => {
 	it('returns proper name', () => {
 		expect(formatVariableName('lastName')).toBe('$lastName');
+	});
+});
+
+describe('formatVariableNameWithType', () => {
+	it('returns proper name', () => {
+		expect(formatVariableNameWithType('foo', 'string')).toBe('$foo (string)');
 	});
 });